@@ -1,7 +1,8 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net5.0</TargetFramework>
+    <Nullable>enable</Nullable>
   </PropertyGroup>
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
@@ -16,9 +17,7 @@
     <None Remove="broker.db" />
     <None Remove="broker.db-shm" />
     <None Remove="broker.db-wal" />
-    <None Remove="cert.pem" />
     <None Remove="fullchain1.pem" />
-    <None Remove="key.pem" />
   </ItemGroup>
 
   <ItemGroup>
@@ -41,6 +40,15 @@
 
   <ItemGroup>
     <ProjectReference Include="..\PainlessMesh\PainlessMesh.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <None Update="cert.pem">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="key.pem">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
   </ItemGroup>
 
 </Project>
