@@ -1,4 +1,7 @@
 ﻿using AppBokerASP.Devices;
+
+using NLog.Fluent;
+
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -40,12 +43,24 @@ namespace AppBokerASP.Database
         public static bool MergeDeviceWithDbData(Device d)
         {
             using var cont = BrokerDbContext;
-            var dbDevice = cont.Devices.FirstOrDefault(x => x.Id == d.Id);
-            if (dbDevice != default)
+            try
             {
-                d.FriendlyName = dbDevice.FriendlyName;
-                return true;
+                var dbDevice = cont.Devices.FirstOrDefault(x => x.Id == d.Id);
+                if (dbDevice != default)
+                {
+                    if (string.IsNullOrWhiteSpace(dbDevice.FriendlyName))
+                        d.FriendlyName = dbDevice.Id.ToString();
+                    else
+                        d.FriendlyName = dbDevice.FriendlyName;
+                    return true;
+                }
             }
+            catch
+            {
+                d.FriendlyName = "";
+                
+            }
+         
             return false;
         }
     }
