@@ -9,17 +9,15 @@ using System.Threading.Tasks;
 
 namespace AppBokerASP.Database
 {
-    public class BrokerDbContext : DbContext
+    public class BrokerDbContextOld : DbContext
     {
         public DbSet<HeaterConfigModel> HeaterConfigs { get; set; }
-        //public DbSet<HeaterConfigModel> HeaterCalibrations { get; set; }
         public DbSet<DeviceModel> Devices { get; set; }
         public DbSet<DeviceMappingModel> DeviceToDeviceMappings { get; set; }
-        //public DbSet<HeaterConfigTemplateModel> HeaterConfigTemplates { get; set; }
 
         protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
         {
-            optionsBuilder.UseSqlite("Data Source=broker.db");
+            optionsBuilder.UseSqlite("Data Source=broker2.db");
         }
         protected override void OnModelCreating(ModelBuilder modelBuilder)
         {
